requirements.txt
streamlit
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-llms-groq
chromadb
sentence-transformers
pymupdf
gtts
python-dotenv
langchain-groq